Get rid of the launcher process
The simterpose process used to start a so called launcher process,
give each command lines to it thru a pipe, and wait for this guy to do
the grunt work. Afterward, that guy was still around, ignored by
simterpose, in charge of just waiting the childs.

Passing the cmdline to the launcher was uselessly complex. Simterpose
knows it as an array, so it had to join and print that array to the
pipe, while the launcher had to read it and split back the array.

Getting simterpose to know which pid got each started process was also
extremely complex: simterpose ptraced the launcher, and when the
launcher ran a fork() (or a variant), simterpose interrupted that
call, retrieved the pid of the granchild directly in the syscall, and
finally restarted the interrupted syscall.

There is actually no need for such a launcher. In the new model,
simterpose (who has the cmdline arguments in the right form for
execvp) starts the child processes directly, and retrieve their pid
from the fork syscall directly.

This seems much more straightforward this way (and seems to lead to
the same result).